requirements.txt
streamlit==1.31.1
google-generativeai
PyPDF2==3.0.6

